% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(filters, structure, latest_by = NULL)
}
\arguments{
\item{filters}{API filters.}

\item{structure}{Structure parameter.}

\item{latest_by}{Retrieves the latest value for a specific metric.
Must be set to a value that is defined in the
structure. (Default: \code{NULL})}
}
\value{
list      Data for the given \code{filters} and \code{structure}.
}
\description{
Extracts paginated data by requesting all of the pages
and combining the results.
}
\details{
For additional information and up-to-date details on arguments
and what they represent, please visit
the \href{https://coronavirus.data.gov.uk/developers-guide}{API documentations}.
}
\examples{
# We would like to download all cases data at `region` level. 
# We start off by defining our `filters` argument:
query_filters <- c(
   "areaType=region"
)

# Next, we define the structure:
query_structure <- list(
    date = "date", 
    name = "areaName", 
    code = "areaCode", 
    daily = "newCasesBySpecimenDate",
    cumulative = "cumCasesBySpecimenDate"
)

# We then pass these arguments to the `get_data` function:
data <- get_data(filters = query_filters, structure = query_structure)
}
