% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_head}
\alias{get_head}
\title{Get Head}
\usage{
get_head(filters, structure)
}
\arguments{
\item{filters}{API filters. See the API documentations for
additional information.}

\item{structure}{Structure parameter. See the API documentations
for additional information.}
}
\value{
list      Request headers.
}
\description{
Request header for the given input arguments (\code{filters},
\code{structure}, and \code{lastest_by}).
}
\examples{
query_filters <- c(
   "areaType=region"
)

query_structure <- list(
    date = "date", 
    name = "areaName", 
    code = "areaCode", 
    daily = "newCasesBySpecimenDate",
    cumulative = "cumCasesBySpecimenDate"
)

headers <- get_head(filters = query_filters, structure = query_structure)

# We can now access header parameters. For instance, to get the 
# timestamp for the latest update, we do as follows:
print(headers$`last-modified`)
}
