% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_options}
\alias{get_options}
\title{Get Options}
\usage{
get_options()
}
\value{
character  API options as prettified JSON.
}
\description{
Provides the options by calling the \code{OPTIONS} method of the API.
}
\examples{
opts <- get_options()
}
