% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_paginated_data}
\alias{get_paginated_data}
\title{Get Paginated Data}
\usage{
get_paginated_data(filters, structure)
}
\arguments{
\item{filters}{API filters.}

\item{structure}{Structure parameter.}
}
\value{
list      Data for the given \code{filters} and \code{structure}.
}
\description{
Iteratively runs the query to download all pages.
}
