% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_release_timestamp}
\alias{get_release_timestamp}
\title{Get Release Timestamp}
\usage{
get_release_timestamp()
}
\value{
closure   Timestamp of the last update.
}
\description{
Produces the website timestamp in GMT.
}
\details{
Added in version 1.2.0

This function supplies the website timestamp - i.e. the time at which the data
were released to the API and by extension the website. Please note that there
will be a difference between this timestamp and the timestamp produced using
the \code{last_update()} function. The latter signifies the time at which the data
were deployed to the database, not the time at which they were released.

It may take up to 60 seconds for the data to be updated when the release timestamp
(\code{get_release_timestamp()}) is updated. This is because the cache must refresh before
the new data becomes available. The API timestamp (\code{last_update()}), however, is only
updated when the cache has been refreshed. This means that you can only be certain that
you are receiving the most up-to-data data when the \code{last_update()} timestamp for your
specific parameters have been updated.

Note: The output is extracted from the header and is accurate to the miliseconds.

Produces the timestamp for the last update in GMT.
}
\examples{
timestamp <- get_release_timestamp()
}
