% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{last_update}
\alias{last_update}
\title{Last Update Timestamp}
\usage{
last_update(filters, structure)
}
\arguments{
\item{filters}{API filters. See the API documentations for
additional information.}

\item{structure}{Structure parameter. See the API documentations
for additional information.}
}
\value{
closure   Timestamp of the last update.
}
\description{
Produces the timestamp for the last update in GMT.
}
\examples{
query_filters <- c(
   "areaType=region"
)

query_structure <- list(
    date = "date", 
    name = "areaName", 
    code = "areaCode", 
    daily = "newCasesBySpecimenDate",
    cumulative = "cumCasesBySpecimenDate"
)

timestamp <- last_update(filters = query_filters, structure = query_structure)
}
